+++ gymnasium/experimental/__init__.py
@@ -1,7 +1,7 @@
 """Root __init__ of the gym experimental wrappers."""
 
 
-from gymnasium.experimental import functional, wrappers
+from gymnasium.experimental import functional
 from gymnasium.experimental.functional import FuncEnv
 from gymnasium.experimental.vector.async_vector_env import AsyncVectorEnv
 from gymnasium.experimental.vector.sync_vector_env import SyncVectorEnv
@@ -12,12 +12,9 @@ __all__ = [
     # Functional
     "FuncEnv",
     "functional",
-    # Wrappers
-    "wrappers",
     # Vector
     "VectorEnv",
     "VectorWrapper",
     "SyncVectorEnv",
     "AsyncVectorEnv",
-    # "vector",
 ]
