+++ gymnasium/experimental/functional_jax_env.py
@@ -11,7 +11,7 @@ import numpy as np
 import gymnasium as gym
 from gymnasium.envs.registration import EnvSpec
 from gymnasium.experimental.functional import ActType, FuncEnv, StateType
-from gymnasium.experimental.wrappers.jax_to_numpy import jax_to_numpy
+from gymnasium.experimental.wrappers.conversion.jax_to_numpy import jax_to_numpy
 from gymnasium.utils import seeding
 from gymnasium.vector.utils import batch_space
